requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
Django==6.0.4
elevenlabs==2.44.0
groq==1.2.0
gTTS==2.5.4
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
packaging==26.1
PyAudio==0.2.14
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
